[SPARK-50977][CORE] Enhance availability of logic performing aggregation of accumulator results
### What changes were proposed in this pull request?

In this PR we introduce a minor refactor that enhances the availability of the functionality used to perform aggregation of accumulator results.

### Why are the changes needed?
These changes make the aggregation logic accessible from other modules which enables various memory and disk optimizations.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

This is a minor refactor. No new code has been added, therefore no new tests are needed.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49618 from costas-db/costas-zarifis_data/refactorMetrics.

Lead-authored-by: Costas Zarifis <costas.zarifis@databricks.com>
